% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{calc_pss}
\alias{calc_pss}
\title{Calculate PSS}
\usage{
calc_pss(f, abs)
}
\arguments{
\item{f}{A named vector of signal intensities per wavelength.}

\item{abs}{Absorbance values of Pr and Pfr.}
}
\value{
A PSS value
}
\description{
Calculate PSS
}
