% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{plot_spectra}
\alias{plot_spectra}
\title{Plot spectrum}
\usage{
plot_spectra(f, xlab = "wavelength", ylab = "Ee (W/(m^2*nm))", ...)
}
\arguments{
\item{f}{Named vector of intensity}

\item{...}{Arguments passed to plot}
}
\description{
Plot spectrum
}

