% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{read_specdat}
\alias{read_specdat}
\title{Read spectral data from Yeti}
\usage{
read_specdat(file, sep = ";", dec = ",")
}
\arguments{
\item{file}{A file name.}
}
\value{
A list containing metadata and spectral data.
}
\description{
Read spectral data from Yeti
}
