% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_functions.R
\name{wavelengthRGB}
\alias{wavelengthRGB}
\title{Create false color from wavelength}
\usage{
wavelengthRGB(wl)
}
\arguments{
\item{wl}{Interger. Wavelength.}
}
\value{
A hexadecimal RGB value.
}
\description{
Create false color from wavelength
}

